+++ src/main/java/de/dm/intellij/liferay/language/osgi/MetaConfigurationReferenceContributor.java
@@ -1,0 +1,29 @@
+package de.dm.intellij.liferay.language.osgi;
+
+import com.intellij.patterns.PlatformPatterns;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiReference;
+import com.intellij.psi.PsiReferenceContributor;
+import com.intellij.psi.PsiReferenceProvider;
+import com.intellij.psi.PsiReferenceRegistrar;
+import com.intellij.psi.filters.position.FilterPattern;
+import com.intellij.util.ProcessingContext;
+import org.jetbrains.annotations.NotNull;
+
+public class MetaConfigurationReferenceContributor extends PsiReferenceContributor {
+
+    @Override
+    public void registerReferenceProviders(@NotNull PsiReferenceRegistrar registrar) {
+        registrar.registerReferenceProvider(
+                PlatformPatterns.psiElement(PsiElement.class).and(new FilterPattern(new MetaConfigurationElementFilter())),
+                new PsiReferenceProvider() {
+                    @Override
+                    public PsiReference @NotNull [] getReferencesByElement(@NotNull PsiElement element, @NotNull ProcessingContext context) {
+                        return new PsiReference[] {
+                                new MetaConfigurationReference(element)
+                        };
+                    }
+                }
+        );
+    }
+}