+++ testdata/de/dm/intellij/liferay/language/osgi/MetaConfigurationReferenceContributorTest/de/dm/action/MyConfigurableAction.java
@@ -1,0 +1,10 @@
+package de.dm.action;
+
+import org.osgi.service.component.annotations.Component;
+
+@Component(
+        configurationPid = "<caret>"
+)
+public class MyConfigurableAction {
+
+}