+++ testdata/de/dm/intellij/liferay/language/osgi/MetaConfigurationReferenceContributorTest/de/dm/action/MyConfigurableReferenceAction.java
@@ -1,0 +1,10 @@
+package de.dm.action;
+
+import org.osgi.service.component.annotations.Component;
+
+@Component(
+        configurationPid = "de.dm.configuration.MyC<caret>onfiguration"
+)
+public class MyConfigurableReferenceAction {
+
+}