+++ testdata/de/dm/intellij/liferay/language/osgi/MetaConfigurationReferenceContributorTest/de/dm/action/MyMultipleConfigurableAction.java
@@ -1,0 +1,13 @@
+package de.dm.action;
+
+import org.osgi.service.component.annotations.Component;
+
+@Component(
+        configurationPid = {
+                "foo",
+                "<caret>"
+        }
+)
+public class MyMultipleConfigurableAction {
+
+}