+++ testdata/de/dm/intellij/liferay/language/osgi/MetaConfigurationReferenceContributorTest/de/dm/action/MyMultipleConfigurableReferenceAction.java
@@ -1,0 +1,13 @@
+package de.dm.action;
+
+import org.osgi.service.component.annotations.Component;
+
+@Component(
+        configurationPid = {
+                "foo",
+                "de.dm.configuration.MyC<caret>onfiguration"
+        }
+)
+public class MyMultipleConfigurableReferenceAction {
+
+}