+++ changelog.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## 0.0.15
 - Feature: Ability to compare theme files with the original files from the parent theme
+- Feature: Resolve configurationPid OSGi properties to configuration interface classes
 
 ## 0.0.14
 - Feature: Resolve Resource Bundles